feat: dispatch built-in calls via registry
