+++ src/assembler/object-layout.ts
@@ -1,0 +1,1 @@
+export const OBJECT_FIELDS_OFFSET = 2;
